réglage d'une incohérence
dans la fonction setPixel

+++ pixel.cpp
@@ -31,7 +31,7 @@ bool pixel::getV(){
 }
 
 
-void pixel::setPixel(double x1, double y1, Color col1, double c1, double d1, bool v1){
+void pixel::setPixel(int x1, int y1, Color col1, double c1, double d1, bool v1){
     x=x1;
     y=y1;
     col=col1;
